Feedback. First agree to rename then pick the names.
Incorporate most of the feedback so far.

The first aim is to decide to rename `int` and `uint`. The second step is to pick the specific names.

I see the second step as usability design, not really bikeshedding, but it is second. A quick-and-dirty usability test could be done. (Although I do like the case for `iptr` and `uptr`.)

TODO: Add usage examples.