Proposed EOF reinstatement.
Precedence: C, C++, Go, and Java all return EOFs as error codes when
the stream ends.  Java includes higher-level APIs that return nulls,
about as close to the current Rust implementation as can be, given the
language differences.

Precedence: Win32 APIs and Windows-based stuff in general tends to
return 'success' and an indicator that it read 0 bytes when such calls
fail.  Well, mostly.  Asynchronous calls are weird, and if you supply
one of the parameters that is commonly null, then Win32's "ReadFile"
will actually start to give you EOFs.

Personally, I have a strong preference for using EOFs a la "the C way"
everywhere.  Rust being sold as a systems programming language makes me
expect this kind of behavior.  Hearing that the IO reforms are trying
to simplify and not implement the union of things between platforms
again implies to me that this will work in the "standard" way.