Updated to reflect comments on earlier RFC.
Added `DropGlue` marker trait, which is defined for a type when it
has drop-glue (meaning that it must be fully populated on scope-based
consumption).
Make the drop-glue function explicit in the language.
Restrict the `T` in `Interior<T>` to `DropGlue` types.
Conversion functions between `T` and `Interior<T>` are now static
functions on `Interior<T>`.
`Interior::from_outer` is made unsafe. Normally, this would violate the
definition of "unsafe", so we also:
Argue for expanding the definition of "unsafe".
More discussion of alternative approaches.