pretty print `Debug` of tuples, tuple structs and enum variants in a single line
Currently, `println!("{:#?}", Some(1));` *pretty* print output as, with newlines and indent:

```
Some(
    1
)
```

I don't think it is **pretty** print: it's weird and ugly, no one write code like this. It is not a preferred code style also.

`Some(1)` itself, without adding newlines and indent, is pretty well.

Tuples were already implemented, in 1.0 stable, to pretty print in a single line, without adding newlines and indents. So we should do the same for tuple structs and enum variants, for consistency.