Variant types and untagged enums
This is something of a two-part RFC, it proposes

* making enum variants first-class types,
* untagged enums (aka unions).

The latter is part of the motivation for the former and relies on the former to
be ergonomic.

In the service of making variant types work, there is some digression into
default type parameters for functions. However, that needs its own RFC to be
spec'ed properly.