Remove `use in match`, add `use in trait`; Add grammar
Remove `use` in `match`. Should be as own RFC, and not willing to defend it personally.
Likewise, removed parts in the how we teach and drawbacks, making those sections less
"if"fy.

Added `trait`, as that's analogous to `impl` and I feel stupid for not doing that initially.

Added grammar changes. Not sure if they're the right changes, but at least they exist.