Fix in-band lifetimes example with backreferences
Backreferences are listed as a "possible extension or alternative" in
RFC #2115, so the examples should not include them. This is further
reinforced by commit c20ea6db254feac3536de7ae8a13bfdee2b07b42, which
appears to have been intended to remove all instances of
backreferences in the examples, but missed these.