Simplify token-based idea.
By focusing less on the proc_macro API, and caring more about how a
library using it would work, we greatly simplify the usage.

Moving to 'only mark a part of the call tokens' removes the need for a
fully-fleged API: once we know how tokens and their properties work we
can hijack that.