Replace "epoch" with "edition"
Future readers interested in the design of editions shouldn't have to
know that we called them "epochs" for a time. They should be able to
find this RFC by searching for "edition", and they shouldn't have to
wonder what an "epoch" is if reading another RFC.

Not changed: the filename of RFC 2052 so that links don't break, and
any usage of the word "epoch" in the "seconds since the epoch" sense.