Add tracking issue for #!feature(async_closure)]
This RFC isn't fully implemented, but tracking bug is closed. 
This PR adds a tracking bug I found for part of the RFC that is not implemented.
There may be other open tracking bugs. This PR is intended to be an improvement,
may not be a complete fix for issue:  https://github.com/rust-lang/rust/issues/66909