Replace `pop` and `pop_before` with `remove_current` in Linked List Cusors
The name "pop" was critized in the discussion thread but despite agreement
it should be changed, no one did. This commit implements basically the
suggestion by Amanieu (but `remove_current` instead of `remove`). There is
some discussion however, over if there should be two methods (differing in
what element is the current one after deletion). This should be discussed
in the tracking issue before stabilizing the feature.