Initial draft of `doctest-name` RFC
This RFC proposes the ability to annotate documentation test code blocks
with metadata of the form `name=NAME`. The given name will be used to
identify the documentation test when it is run by the test runner, in
addition to the current information of test binary, module, and line
number.

For example, this documention test:

    ```name=arithmetic
    assert_eq!(2 + 2, 4);
    ```

Assuming that it is in crate `foo`, in file `src/lib.rs`, in module
`bar`, on line 37, will produce this output when run:

       Doc-tests foo

    running 1 test
    test src/lib.rs - bar::arithmetic (line 37) ... ok

    test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out