Remove the stability system and `TransmuteInto` (for now).
The stability system fails to match the usual rules of stability:
  https://github.com/rust-lang/rfcs/pull/2981#issuecomment-710033947
  https://github.com/rust-lang/rfcs/pull/2981#issuecomment-710048751

Per the rationale here:
  https://github.com/rust-lang/rfcs/pull/2981#issuecomment-710263596
this commit merely removes the stability system, rather than
attempting to replace it. It's hard to craft a stability system
that isn't overly-prescriptive.

I anticipate a not-insignificant amount of design work will be
required to overcome the pitfalls I describe in that comment. It
might very well happen, but it will take time.

Separately, this commit removes `TransmutableInto` (for now).

`TransmuteInto` was necessitated by the `PromiseTransmutableFrom`
trait. Without that trait, it's just an ergonomic nicety. It *is*
nice to have and may very well make a reappearance, but in this
time of flux, it's just an extra trait that requires labor to keep
it up-to-date with the current state of the RFC.