Add some notes to help avoid conflicts of interest
Hopefully this will help anyone who might otherwise be placed in a
position of being expected to approve/reject a target due to employment
or other legal obligations. People in such a position may be able to
point to this language to protect themselves.

This language may additionally help protect Rust team members from
causes of action brought by parties with a vested interest in the status
of a target. (There may, in the future, be *substantial* business
interests depending on the support status of specific targets.)