Make requirements about unwanted messages more strict
Even a well-intentioned comment on a PR stating that it breaks a tier
2/3 target can still derail the PR, and developers may not be aware that
they don't need to act on such messages. Drop the exception and tighten
the language.

Add a note that automated messages or notifications *outside* the PR
thread are not OK either, lest someone generate emails to PR authors
instead, or generate issues elsewhere that `@` the PR author.

Add a note that backlinks (such as those generated by GitHub whenever
linking to a PR) are acceptable within reason. Otherwise, target
developers on a separate repository could not even link to a PR as part
of coordination.