Potential Target Policy
This RFC introduces a policy for targets not yet ready to merge into
Rust, to allow for initial coordination and consensus-seeking.

Rust teams or team members sometimes have informal discussions with the
developers of various potential targets, to coordinate and build
consensus on the requirements needed to later upstream the target, and
to determine properties of the target such as the ABI and target name.
However, such informal discussions do not typically get recorded in any
official location, and may not be coordinated among the broader Rust
project.

This policy (the "Potential Target Policy") allows for the formal
evaluation of a potential target for Rust, and the acceptance of such a
target as a "Potential Target". It has a lower threshold than even the
"Tier 3" target requirements, while still serving to start a
conversation and reach some initial consensus on compatibility.

For example, an approval of a Potential Target may nail down the target
name (the "triple"); this helps avoid the situation where other tools
begin using a name and that name becomes more difficult to change. An
approval of a Potential Target may additionally nail down the sizes and
alignments of various types, and how those types are passed in the C
ABI, to allow coordination among other tools generating code for the
target.

There are many potential names we could use for such targets. I chose
"Potential Target" over names like "Future Target" to avoid any
implication of upcoming/impending inclusion of the target. "Add xyz as
a Potential Target" seems like an accurate description of what this
process will enable.