Clarify naked function calling convention reqs
Although we don't have `extern "custom"` anymore, `unsafe` can
pragmatically fill this roll. Marking a function as `unsafe` implies
that the function cannot be safely called unless additional criteria are
met (as documented in the safety section, per standard convention). This
wording clarifies the RFC to point to this well-established convention.