Expand on the guide-level explanation
This is pulled from https://github.com/rust-lang/rfcs/pull/3243#issuecomment-1064664638

This shifted the guide to starting from the consumer of a namespace, as
that is the majority case, and then moving into the role of a crate
owner.

This leaves out details on feature flags to not duplicate a future "why
not feature flags" section.