Updaet motivation for `::` semantics
This was meant to pull in the guide-level explanation's reasons for this
feature but they were mostly covered.  Instead, I ended up re-wording
parts because of the semantics of `::` changes slightly the problem this
solves.  This is no longer a general project solution but specific to
when everything makes sense in the same API.  These updates are intended
to convey that.