Revise options for `#[refine]`
Emphasize the possibility of continuing to use `#[refine]` in future
editions, and list out all the options under Unresolved Questions.
Include the attribute in examples.

Simplify the usage of `#[refine]` in existing editions so that it is
needed to refine any aspect of an API, instead of new features being
refine-able by default and old ones needing the attribute.

Move any discussion of *not* doing a soft transition to Unresolved
Questions to simplify the Reference Level Explanation a bit.