Apply suggestions from code review
Wesley pointed out that the text referred to "larger initiatives related to the language" is sort of weird given that the scope of this document is about the compiler.

Reasonable people can disagree here (as evidenced by my response to Wesley), but for this text, I'm going to go ahead and revise the language to follow Wesley's suggestion. Its not a detail that's worth getting hung up over, since this is a laundry list of things that are explicitly *not* the sole domain of the team lead, that's all its meant to be, and its not meant to be an exhaustive list, for that matter.

Co-authored-by: Wesley Wiser <wwiser@gmail.com>