Rewrite to use an attribute instead of a trait
Responding to feedback that a trait/use of the type system is too
invasive. This version has the compiler infer for generic functions
the list of generic types which may be dropped, and propagates that to
dependent crates. Users are no longer expected to explicitly annotate
their generic functions, nor do they need to `#[derive(Destruct)]`.