Disable the playground run button by default for code blocks
Most example code in RFCs is not written with executability in mind, so having the "Run" button available by default on all RFCs seems a bit misleading.
This now disables the playground for the whole book (which also removes the "hidden" code mdbook inserts). It thus only leaves the copy button.