Add RFC for lifetime capture rules 2024
The lifetime capture rules discussed in this RFC were the subject of a
T-lang design meeting on 2023-07-26.  The consensus of that meeting
was to put these rules into effect for the Rust 2024 edition and for
all RPIT-like features stabilized ahead of the Rust 2024 edition.

The purpose of this RFC is to memorialize the consensus of that
meeting.