Fix TAIT row in Appendix B matrix
In Appendix B, we have a matrix of capturing effects.  For TAIT, we
had correctly marked the capturing of lifetime parameters from the
outer impl as being not applicable in Rust 2021, but we had
incorrectly marked this as being "yes" in Rust 2024.  It's of course
not applicable in both editions.  Let's fix that.