Rework some of the text around "missing an edition".
Some people have expressed different thoughts on how this scenario
should be addressed. However, it seems like it will be a very
situational event, and trying to predict exactly what we should do seems
over-prescriptive. The RFC leaves the decision up to the Leadership
Council who is expected to take the different circumstances and desires
into consideration.