Clarify outlives rules on opaques with specified bounds
There is some subtlety in describing the mechanics of the outlives
rules for `impl Trait` opaque types.  In our description of these
rules for opaque types that include a specified outlives bound, we had
some words that were less clear than we had intended.  Let's change
those words around to make the meaning more clear.

Specifically, rather saying that "it" outlives the specified lifetime,
what we mean exactly is that the lifetime that's substituted for the
specified lifetime parameter must outlive that other lifetime.

Note that none of this language is normative.  These rules have all
been specified in earlier RFCs.