Lowercase internal link identifiers
Unfortunately, some Markdown tools care that the internal link
identifiers are the same case as the headings with which they are
associated, some tools care that the identifiers are always lower
case, and some tools use case insensitive matching for these
identifiers.

There is no way to satisfy all of these tools simultaneously.
However, the tools we care about at the moment either use case
insensitive matching or care that the identifiers are lower case, so
let's convert all of the identifiers to lower case.