Say "into earlier editions" rather than into Rust 2021
In this RFC, we're specifying the capture rules for new RPIT-like
`impl Trait` features that may be introduced ahead of the Rust 2024
edition.  We had described those as features "introduced into the Rust
2021 edition", but actually, new features are introduced where
possible into all editions, so it's more correct to describe the
features as being "introduced into earlier editions".  Let's make that
change.

(Thanks to @tmandry for the suggestion.)