Add forward link and footnote about `Captures` trait
We never added the `Captures` trait to the standard library, and it's
unlikely that we ever will due to this RFC, so there is no one single
way to define it.  However, different people may have different
expectations about it based on how those people have seen it defined
in the past.

Let's add a forward reference in the first place we use `Captures` to
the section on the `Captures` trick where we define the trait, and
let's add a footnote in that section that notes the existence of other
plausible implementations.

(Thanks to @spastorino for pointing out that `rustc` internally
currently uses a different definition of this trait.)