Move footnote reference outside of colon
We placed the reference to our footnote about how `Captures` can be
defined in various ways inside of the colon that leads into the
example.  Since this footnote is really about the example itself
rather than about the text introducing the example, let's put the
reference to the footnote outside of the colon.  This looks a bit
better typographically anyway.