Move forward `Captures` links to footnotes
We use the `Captures` trait for exposition before formally introducing
it in a later section.  In an earlier commit, we had added some text
in parentheses to explain this.

Rather than breaking up the flow with parenthesized text, let's use
footnotes for this.

In the second place that we do this, we're using `Captures` only for
exposition; real code would not need to use the trick here.  We'll
mention that in the footnote as well.