Improve return type annotation in `async fn` examples
In two examples involving `async fn`, we annotate the returned opaque
type in comments.  In these comments, we had notated the trait bounds,
but we had not included the keyword `impl`.  Let's include that
keyword for greater clarity.

We'll also add backticks around the annotated return type in the
second example.  These were already present in the first example.