Add caveat on `Captures<'a> + Captures<'b>` in footnote
In a footnote, we describe how there are various ways that `Captures`
can be defined.  However, we should point out that `Captures<'a> +
Captures<'b>` is not equivalent to `Captures<(&'a (), &'b ())>` as
lifetimes do not participate in trait selection in Rust.  This is in
fact the main reason that we defined `Captures` to accept a type
parameter rather than a lifetime parameter in this document.  Let's
include this caveat in the footnote.

(Thanks to @compiler-errors for reminding everyone of this.)