Add example and mention of `std::iter::from_fn`
In the motivation, we enumerate some ways that people can create an
iterator.  One way we didn't demonstrate was the use of
`std::iter::from_fn`.  Let's do that, and let's add a mention of this
in the section about the ever-present option of doing nothing.

(Thanks to @estebank for pointing this out.)