Clarify section on error handling
The section on error handling discussed very specifically the behavior
with `Option` and `Result` in a way that may have made the general
case less clear.  E.g., it did not discuss the handling of
`ControlFlow`, leaving unclear whether arbitrary types that implement
the `Try` trait could be used.  It also did not discuss whether
`From::from` was called as usual in case of short-circuiting.

Let's simplify this section a bit by first discussing the desugaring,
discussing the non-effect on the required type of trailing elements or
arguments to `return` in the block, and discussing the effect on the
type of the arguments required by `yield`.

Notably, we won't discuss here the details of how `None` is returned
to end iteration.  This behavior is the same as when `?` is not used,
and the desugaring we describe fills in all of the needed details.

(Thanks to @scottmcm for pointing out the ambiguous nature of this
section.)