Add normative example on `for<..>` capturing
Under the rules of this RFC, we automatically capture all lifetime
parameters in scope of an RPIT-like `impl Trait` opaque type.

This includes lifetimes that are in scope due to `for<..>` binders for
higher ranked trait bounds (HRTBs).  However, we did not describe this
explicitly and we did not have an example of this.  Let's describe
this and add an example.

(Thanks to @aliemjay for pointing this out.)