Deemphasize TAIT a bit
In the design meeting for the 2024 lifetime capture rules, we
presented TAIT as being the intended solution to the problem of
overcapturing.  By "intended", we simply meant that it is known to be
a full solution, and since it has accepted RFCs and a consensus to
stabilize it in some form, it's more *intended* than other solutions
lacking those things.

However, as applied to the rules in this RFC, the only thing that's
important is that *some* solution is stabilized that allows for
existing code to continue to be expressed in some manner.  The RFC had
already said this (e.g., "...is contingent on the stabilization of
TAIT or some other solution..."), but it was a bit obscured by the
emphasis on TAIT.

Let's deemphasize TAIT and describe the landscape of solutions more
fully so as to keep the focus on the new capture rules.  We'll move
the discussion of TAIT out of the normative section entirely and purge
the word "intended" from the document.

(Thanks to @tmandry for pointing this out.)