Remove language about an experiment
Some of the language in this RFC implied that this is an experiment
and that another RFC would be required.  This isn't how we want to
frame things.  Another RFC could always be required, but it's also
possible that stabilization could occur on the basis of this RFC.

This is a normal RFC.  Let's remove all language about this being an
experiment.