fix: Switch from `pub` to `public`
Since we switched from warn-by-default to allow-by-default,
the need to set `pub` on dependencies with an unsupported MSRV is much
lower.
It'd help with people wanting to enable this now, especially for core
packages that are more likely to have lower MSRVs (but also less likely
to have dependencies).

So if we drop the requirement for allowing `pub` on old MSRVs, we can
re-visit using the name `public`.