Add RFC for macro fragment specifiers edition policy
Recently T-lang discussed the issue of macro matcher fragment
specifiers, e.g. `expr`, falling out of sync with the underlying
grammar of Rust.  The consensus of the meeting was that it would be
proper to set out a policy for how these divergences might be
reconciled over edition boundaries.

The purpose of this RFC is to memorialize the resulting consensus.