Specify when specifiers are added and to what editions
In the meeting, we did not discuss whether the new fragment specifiers
that preserve the old behavior should be added to all editions or only
to the new edition, and accordingly, we did not discuss when these
should be added.

In keeping with our general policy of working to minimize differences
between editions, it would seem to make sense to add these fragment
specifiers to all editions.  Given that, to provide people with the
most time possible to prepare for the new edition, it would make sense
to add these as early as possible and no later than when the new
edition is released.

Let's update the RFC to specify that.

(Thanks to Ralf Jung for asking about this.)