Add alternative of new edition behavior in all editions
One thing we could do, but probably won't, is to add a fragment
specifier to all editions with the behavior of the specifier in the
new edition.  E.g., when releasing Rust 2024, we could add an
`expr2024` fragment specifier to all editions.

Let's add a section describing this and why it probably won't happen.

(Thanks to deltragon for bringing this up.)