Switch from `expr2021` to `expr_2021`
We had originally specified that, when a better semantically
meaningful name could not be found, the new fragment specifier would
be named by using the existing name with the identifier of the current
edition added directly as a suffix, e.g. `expr2021`.  However,
reviewing RFC 430 and existing practice, we should probably instead
use an underscore to separate the edition identifier,
e.g. `expr_2021`.

Let's update the RFC to specify that the underscore separator will be
used, and let's add a section that discusses the other possible
alternative choices.

(Thanks to Nilstrieb for raising this point.)