Add open question about Rust 2015 and Rust 2018
In RFC 3101 we reserved in Rust 2021 prefixed identifiers such as
`prefix#ident`.  For this reason, we can make `gen` blocks available
in Rust 2021 using `k#gen` as was anticipated in the (currently
pending) RFC 3098.

It's less clear what to do about Rust 2015 and Rust 2018, however, so
let's mark this as an open question.

(Thanks to tmandry for raising this point.)