Add RFC 3617 for precise capturing
To fully stabilize, in Rust 2024, the Lifetime Capture Rules 2024 that
we accepted in RFC 3498, we need to stabilize some means of precise
capturing.  This RFC provides that means.

We discussed this feature, the need for it, and the syntax for it in
the T-lang planning meeting on 2024-04-03.  The document here follows
from that discussion.