Rework the syntax specification
For the formal syntax, we had used the existing `GenericParams`
production.  However, that production isn't exactly appropriate for
this case.  What's needed here is essentially a generic argument list
that only accepts generic parameters as elements.

Since none of the other existing productions provide this, we'll
define our own.

(Thanks to @kennytm for pointing this out.)