Extend discussion on `use<..> impl Trait` syntax
We had earlier written up a section on `use<..> impl Trait` syntax
that mostly focused on why we had not adopted it.  We didn't spend
much text on why it's appealing, probably because we are in fact
sympathetic to it and consider the reasons that it's appealing
obvious.

Still, we should write all those reasons down.  Let's extend the
section on this syntax with the best possible argument in favor.

We also see more clearly now the fundamental intuitive tension behind
this syntax and `impl use<..> Trait`, so let's write that down too.

Finally, let's describe the historical and other factors that led to
picking one syntax over the other.

(Thanks to tmandry for suggesting the `use<..> impl Trait` syntax and
many of the arguments in favor of it.)