Improve example on avoiding HR lifetime capture
We have an example showing how to avoid capturing a higher ranked
lifetime in a nested opaque type so that the code can be migrated to
Rust 2024.  However, because we didn't parameterize the trait in the
example with a lifetime, the code could be migrated by just removing
the `for<..>` binder.  This makes the example weaker than it could be,
so let's strengthen it by parameterizing the trait.

(Thanks to aliemjay for raising this point.)