Extend and improve reference desugarings
We had included a reference desugaring of `use<..>` in RPIT using
ATPIT.  Let's add a similar desugaring for `use<..>` in RPITIT.

In doing this, we'll make some changes to the RPIT desugaring so that
it better parallels the RPITIT one.  In particular, we want to
turbofish all generic parameters for clarity, and we want to eliminate
all function arguments for conciseness.  Doing this means that all of
the lifetimes become early bound.  This seems fine, since it's rather
orthogonal to the semantics we're trying to demonstrate here.

We also want to demonstrate using const generics in the hidden type.
We could do this using arrays, e.g. `[(); N]`, but it seems more clear
to just define a type constructor that uses all of the generics, so
we'll sacrifice a bit of conciseness to do it that way.