Simplify RPITIT desugaring
We had added an RPITIT desugaring that was rather complicated.  This
complication was due to trying to explain what it would mean to not
capture generic parameters that are part of the trait header.

However, it's likely that there's no good way to express that semantic
in the surface syntax.  Let's instead simplify the desugaring and make
a note of its limitations.