Prepare RFC 3606 to be merged
The FCP has completed on RFC 3606, so let's prepare it to be merged.

First, we'll shorten the name of the feature flag a bit; this should
still be unambiguous.

Second, we're going to remove the graphic from the summary.  While it
may be illustrative, the text and the other examples seem clear enough
without it, and its benefits have to be weighed against the fact that
we want the content in this repository to be easily editable and
freestanding.  Pulling in an SVG file from an outside host pulls
against that.  If we come to think the graphic is critical, we could
always add it back in a separate PR that would add an editable version
of this SVG file into the repository itself.

Third, let's make the H1 title of the document a bit more clear.