Address feedback and questions
As this RFC was reviewed in the GitHub thread, many alternatives were
proposed and questions raised.  As those of us who were a bit too
close to it were cursed with knowledge, the rationale for rejecting
these alternatives were not fully articulated and not all of these
questions were clearly answered.

Let's better document these alternatives, the rationale for rejecting
each, and the answers to various known questions.

(Thanks to GoldsteinE, madsmtm, kennytm, samih, and tmccombs for
raising these alternatives and questions.)