Clarify about `unsafe_code` and edition migration
When people migrate to the new edition, if they have turned up the
severity of the `unsafe_code` lint and they have `extern` blocks that
need to be marked `unsafe`, they will see this lint as intended.
Let's make a note of that.

(Thanks to kennytm for raising this point.)