Fix optionality of `safe`/`unsafe` in guide section
The 2023-10-11 consensus from lang on how to move forward included
that the `safe` and `unsafe` keywords would be optional within an
`unsafe extern` block.

The reference-level section correctly followed this consensus, but the
guide-level section did not, and suggested that annotating items with
these keywords was required.  Let's fix that.