Add the numbered match ergonomics rules adopted by lang
In the lang design meeting on 2024-06-26, we adopted a slate of five
rules for match ergonomics.  We later amended *Rule 4* in our meeting
on 2024-07-03.

These five rules, as amended, form the normative description of what
the lang team means for the rules to be.  Therefore, in the
reference-level explanation, let's include these rules verbatim and
note that the remaining sections describe these rules further.