Update summary per 2025-03-19 lang design meeting
We had consensus in the meeting to do `PhantomNotFreeze` and that
`PhantomData<T>` should be `Freeze` only if `T: Freeze`.  Let's note
those in the summary.

On the name, the meeting leaned heavily toward keeping the name
`Freeze`, and definitely away from `ShallowImmutable`, so let's remove
the reference to the latter.