Remove the open questions
Let's not leave open the question about how the marker trait should
work in terms of `Send`.  We'll specify that in the summary for now,
but probably that should be addressed elsewhere in the RFC also.

The question of what the standard library might do in the future in
terms of making guarantees can be framed as a future possibility.