Elaborate on motivation in template
The template had a very short explanation of the motivation, which
probably led many people into believing that the motivation can be
rather short, compared to the explanations.

But really, the motivation is extremely important and should often
contain a lot of information. This makes those expectations a lot more
explicit. It also explicitly asks for use cases to be listed out, which
can then be referenced in the rest of the RFC to make it clearer how it
solves the problem.